Bug 1853012 - Update libcubeb to revision f9c118d. r=cubeb-reviewers,chunmin
Differential Revision: https://phabricator.services.mozilla.com/D188348

+++ media/libcubeb/src/cubeb_opensl.cpp
@@ -1852,7 +1852,7 @@ opensl_stream_get_position(cubeb_stream * stm, uint64_t * position)
     stm->lastPosition = msec;
   }
 
-  uint32_t samplerate = stm->user_output_rate;
+  uint64_t samplerate = stm->user_output_rate;
   uint32_t output_latency = stm->output_latency_ms;
 
   pthread_mutex_lock(&stm->mutex);
